Split all client side pagees into a client side view and a serverside page to support static page titles
