+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "js/ts.tsdk.path": "node_modules/typescript/lib"
+}
