+++ src/app/projects/page.tsx
@@ -1,46 +1,11 @@
-"use client";
+import { Metadata } from "next";
+import View from "./view";
 
-import GenericTableView from "@/src/components/GenericTableView/GenericTableView";
-import { createCellRenderer } from "@/src/utils/cellRenderers";
-
-const headers = ["id", "Name", "Staff", "Status", "Project URL", "Last Contact"];
-const cellRenderer = createCellRenderer({ editPath: "/projects/edit" });
+export const metadata: Metadata = {
+  title: 'Projects',
+};
 
 function Page() {
-  return (
-    <GenericTableView
-      headers={headers}
-      cellRenderer={cellRenderer}
-      sortableColumns={{
-        Name: "name",
-        Staff: "staff1",
-        Status: "status",
-        "Last Contact": { column: "last_contact", default: "desc" },
-      }}
-      query={async (client, opts, searchQuery) => {
-        const queryObj = searchQuery ? { name: `ilike.${searchQuery}` } : undefined;
-        const result = await client.getProjects({
-          ...opts,
-          ...(queryObj && { query: queryObj }),
-        });
-
-        const data = result.data.map((project) => [
-          project.id,
-          project.name,
-          project.staff1?.name ?? project.staff1?.email1 ?? "",
-          project.status ?? "",
-          project.url ?? "",
-          project.last_contact ?? "",
-        ]);
-
-        return { data, totalCount: result.totalCount };
-      }}
-      queryLabel="Search by Project Name"
-      timeColumn="Last Contact"
-      linkColumn="Project URL"
-      unauthenticatedMessage="You must be logged in to view projects."
-    />
-  );
+  return <View />
 }
-
 export default Page;
